Refactors to separate downloading capabilities from the rest (#1)
* Refactors to separate downloading capabilities from the rest
* import errors
* circumvents NetworkX lack of algorithm
