+++ tests/test_osmnx.py
@@ -5,14 +5,18 @@
 ################################################################################
 
 import matplotlib as mpl
-mpl.use('Agg') #use agg backend so you don't need a display on travis-ci
+import warnings
+
+mpl.use('Agg')  # use agg backend so you don't need a display on travis-ci
 
 # remove the .temp folder if it already exists so we start fresh with tests
 import os, shutil
+
 if os.path.exists('.temp'):
     shutil.rmtree('.temp')
 
 import osmnx as ox
+from networkx.exception import NetworkXNotImplemented
 
 # configure OSMnx
 ox.config(log_console=True, log_file=True, use_cache=True,
@@ -21,7 +25,6 @@ ox.config(log_console=True, log_file=True, use_cache=True,
 
 
 def test_imports():
-
     # test all of OSMnx's module imports
     import ast
     import datetime
@@ -63,7 +66,6 @@ def test_imports():
 
 
 def test_logging():
-
     # test OSMnx's logger
     import logging as lg
     ox.log('test a fake debug', level=lg.DEBUG)
@@ -75,7 +77,6 @@ def test_logging():
 
 
 def test_geometry_coords_rounding():
-
     # test the rounding of geometry coordinates
     from shapely.geometry import Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon
 
@@ -103,10 +104,9 @@ def test_geometry_coords_rounding():
 
 
 def test_gdf_shapefiles():
-
     # test loading spatial boundaries, saving as shapefile, and plotting
     city = ox.gdf_from_place('Manhattan, New York City, New York, USA')
-    city_projected = ox.project_gdf(city, to_crs={'init':'epsg:3395'})
+    city_projected = ox.project_gdf(city, to_crs={'init': 'epsg:3395'})
     ox.save_gdf_shapefile(city_projected)
 
     city = ox.gdf_from_place('Manhattan, New York City, New York, USA', buffer_dist=100)
@@ -114,7 +114,6 @@ def test_gdf_shapefiles():
 
 
 def test_graph_from_file():
-
     # test loading a graph from a local .osm file
     import bz2, tempfile
 
@@ -140,7 +139,6 @@ def test_graph_from_file():
 
 
 def test_network_saving_loading():
-
     # save/load graph as shapefile and graphml file
     G = ox.graph_from_place('Piedmont, California, USA')
     G_projected = ox.project_graph(G)
@@ -168,7 +166,6 @@ def test_network_saving_loading():
 
 
 def test_get_network_methods():
-
     from shapely import wkt
 
     # graph from bounding box
@@ -183,14 +180,16 @@ def test_get_network_methods():
     G3 = ox.graph_from_point(location_point, distance=500, distance_type='network')
 
     # graph from address
-    G4 = ox.graph_from_address(address='350 5th Ave, New York, NY', distance=1000, distance_type='network', network_type='bike')
+    G4 = ox.graph_from_address(address='350 5th Ave, New York, NY', distance=1000, distance_type='network',
+                               network_type='bike')
 
     # graph from list of places
-    places = ['Los Altos, California, USA', {'city':'Los Altos Hills', 'state':'California'}, 'Loyola, California']
+    places = ['Los Altos, California, USA', {'city': 'Los Altos Hills', 'state': 'California'}, 'Loyola, California']
     G5 = ox.graph_from_place(places, network_type='all', clean_periphery=False)
 
     # graph from polygon
-    polygon = wkt.loads('POLYGON ((-122.418083 37.754154, -122.418082 37.766028, -122.410909 37.766028, -122.410908 37.754154, -122.418083 37.754154))')
+    polygon = wkt.loads(
+        'POLYGON ((-122.418083 37.754154, -122.418082 37.766028, -122.410909 37.766028, -122.410908 37.754154, -122.418083 37.754154))')
     G6 = ox.graph_from_polygon(polygon, network_type='walk')
 
     # test custom query filter
@@ -203,7 +202,6 @@ def test_get_network_methods():
 
 
 def test_stats():
-
     # create graph, add bearings, project it
     location_point = (37.791427, -122.410018)
     G = ox.graph_from_point(location_point, distance=500, distance_type='network')
@@ -214,11 +212,15 @@ def test_stats():
     stats1 = ox.basic_stats(G)
     stats2 = ox.basic_stats(G, area=1000)
     stats3 = ox.basic_stats(G_proj, area=1000, clean_intersects=True, tolerance=15, circuity_dist='euclidean')
-    stats4 = ox.extended_stats(G, connectivity=True, anc=True, ecc=True, bc=True, cc=True)
+
+    try:
+        stats4 = ox.extended_stats(G, connectivity=True, anc=True, ecc=True, bc=True, cc=True)
+    except NetworkXNotImplemented as e:
+        warnings.warn("Testing coordinates results in a MultiDigraph, and extended stats are not available for it")
+        warnings.warn(e.args)
 
 
 def test_plots():
-
     G = ox.graph_from_place('Piedmont, California, USA', network_type='drive', simplify=False)
     G2 = ox.simplify_graph(G, strict=False)
 
@@ -248,7 +250,6 @@ def test_plots():
 
 
 def test_routing_folium():
-
     # calculate shortest path and plot as static image and leaflet web map
     import networkx as nx
     G = ox.graph_from_address('398 N. Sicily Pl., Chandler, Arizona', distance=800, network_type='drive')
@@ -272,7 +273,6 @@ def test_routing_folium():
 
 
 def test_nearest_edge():
-
     # test in closest edge section
     sheik_sayed_dubai = [25.09, 25.06, 55.16, 55.11]
     location_coordinates = (25.071764, 55.138978)
@@ -281,7 +281,6 @@ def test_nearest_edge():
 
 
 def test_nearest_edges():
-
     from pyproj import Proj
 
     # test in closest edge section
@@ -291,7 +290,7 @@ def test_nearest_edges():
 
     # Unprojected
     ne1 = ox.get_nearest_edges(G, X=[location_coordinates[1], location_coordinates[1]],
-                                  Y=[location_coordinates[0], location_coordinates[0]], method='balltree', dist=0.0001)
+                               Y=[location_coordinates[0], location_coordinates[0]], method='balltree', dist=0.0001)
 
     # Projected
     G2 = ox.project_graph(G)
@@ -299,7 +298,7 @@ def test_nearest_edges():
 
     projected_point = crs(location_coordinates[1], location_coordinates[0])
     ne2 = ox.get_nearest_edges(G2, X=[projected_point[0], projected_point[0]],
-                                   Y=[projected_point[1], projected_point[1]], method='kdtree', dist=10)
+                               Y=[projected_point[1], projected_point[1]], method='kdtree', dist=10)
     assert (ne1 == ne2).all()
 
 
@@ -320,14 +319,14 @@ def test_footprints():
         new_river_head_responses = [json.load(read_file)]
     new_river_head_gdf = ox.create_footprints_gdf(responses=new_river_head_responses)
     assert 665593284 in new_river_head_gdf.index
-    assert new_river_head_gdf.loc[9246394]['geometry'].type=='Polygon'
-    assert len(new_river_head_gdf.loc[9246394,'geometry'].interiors)==2
+    assert new_river_head_gdf.loc[9246394]['geometry'].type == 'Polygon'
+    assert len(new_river_head_gdf.loc[9246394, 'geometry'].interiors) == 2
 
     # clapham_common.json contains a relation with 5 outer rings and 1 inner ring. One of the outer rings is a chain of open ways
     with open("tests/input_data/clapham_common.json", "r") as read_file:
         clapham_common_responses = [json.load(read_file)]
     clapham_common_gdf = ox.create_footprints_gdf(footprint_type='leisure', responses=clapham_common_responses)
-    assert clapham_common_gdf.loc[1290065]['geometry'].type=='MultiPolygon'
+    assert clapham_common_gdf.loc[1290065]['geometry'].type == 'MultiPolygon'
 
     # relation_no_outer.json contains a relation with 0 outer rings and 1 inner ring
     with open("tests/input_data/relation_no_outer.json", "r") as read_file:
@@ -354,8 +353,8 @@ def test_footprints():
 
     gdf = ox.footprints_from_place(place='kusatsu, shiga, japan', which_result=2)
 
-def test_pois():
 
+def test_pois():
     import pytest
     # download all points of interests from place
     gdf = ox.pois_from_place(place='Kamppi, Helsinki, Finland')
@@ -382,7 +381,6 @@ def test_pois():
 
 
 def test_nominatim():
-
     import pytest
     from collections import OrderedDict
 
@@ -392,13 +390,13 @@ def test_nominatim():
 
     # Bad Address - should return an empty response
     params['q'] = "AAAAAAAAAAA"
-    response_json = ox.nominatim_request(params = params,
-                                         type = "search")
+    response_json = ox.nominatim_request(params=params,
+                                         type="search")
 
     # Good Address - should return a valid response with a valid osm_id
     params['q'] = "Newcastle A186 Westgate Rd"
-    response_json = ox.nominatim_request(params = params,
-                                         type = "search")
+    response_json = ox.nominatim_request(params=params,
+                                         type="search")
 
     # Lookup
     params = OrderedDict()
@@ -406,21 +404,21 @@ def test_nominatim():
     params['address_details'] = 0
     params['osm_ids'] = "W68876073"
 
-    response_json = ox.nominatim_request(params = params,
-                                         type = "lookup")
+    response_json = ox.nominatim_request(params=params,
+                                         type="lookup")
 
     # Invalid nominatim query type
     with pytest.raises(ValueError):
         response_json = ox.nominatim_request(
-                            params = params,
-                            type = "transfer")
+            params=params,
+            type="transfer")
 
     # Searching on public nominatim should work even if a key was provided
     ox.config(
         nominatim_key="NOT_A_KEY"
     )
-    response_json = ox.nominatim_request(params = params,
-                                         type = "search")
+    response_json = ox.nominatim_request(params=params,
+                                         type="search")
 
     # Test changing the endpoint. It should fail because we didn't provide a valid key
     ox.config(
@@ -441,7 +439,6 @@ def test_osm_xml_output():
 
 
 def test_overpass():
-
     import pytest
 
     # Test changing the endpoint. This should fail because we didn't provide a valid endpoint
